use unsigned types in protocols & major robustness improvement of configuration classes & byte processing

+++ dragonite-forwarder/src/main/java/com/vecsight/dragonite/forwarder/network/server/ForwarderServer.java
@@ -28,7 +28,7 @@ public class ForwarderServer {
 
     private final int forwardingPort;
 
-    private final short limitMbps;
+    private final int limitMbps;
 
     private final String welcomeMessage;
